+++ eventum/app/models/parameters/log.py
@@ -14,6 +14,10 @@ class LogParameters(BaseModel, extra='forbid', frozen=True):
         default='info'
         Logging level.
 
+    third_party_level : Literal['debug', 'info', 'warning', 'error',\
+        'critical'], default='warning'
+        Logging level of third-party libraries.
+
     format : Literal['plain', 'json'], default='plain'
         Logging format.
 
@@ -26,6 +30,13 @@ class LogParameters(BaseModel, extra='forbid', frozen=True):
     """
 
     level: Literal['debug', 'info', 'warning', 'error', 'critical'] = 'info'
+    third_party_level: Literal[
+        'debug',
+        'info',
+        'warning',
+        'error',
+        'critical',
+    ] = 'warning'
     format: Literal['plain', 'json'] = 'plain'
     max_bytes: int = Field(default=10 * 1024 * 1024, ge=1024)  # 10MiB
     backups: int = Field(default=5, ge=1)